+++ backend/controller/cronjobs/cronjobs.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/benbjohnson/clock"
 
-	"github.com/TBD54566975/ftl/backend/controller/cronjobs/dal"
+	"github.com/TBD54566975/ftl/backend/controller/cronjobs/internal/dal"
 	parentdal "github.com/TBD54566975/ftl/backend/controller/dal"
 	encryptionsvc "github.com/TBD54566975/ftl/backend/controller/encryption"
 	"github.com/TBD54566975/ftl/backend/controller/encryption/api"
