+++ sqlc.yaml
@@ -6,7 +6,7 @@ sql:
       - backend/controller/dal/internal/sql/queries.sql
       - backend/controller/dal/internal/sql/async_queries.sql
       # FIXME: Until we fully decouple cron from the controller, we need to include the cron queries here
-      - backend/controller/cronjobs/dal/internal/sql/queries.sql
+      - backend/controller/cronjobs/internal/sql/queries.sql
       # Some of the timeline entries happen within a controller transaction, so we need to include them here
       - backend/controller/timeline/internal/sql/deployment_queries.sql
     schema: "backend/controller/sql/schema"
@@ -154,12 +154,12 @@ sql:
       - postgresql-no-seq-scan
   - <<: *daldir
     queries:
-      - backend/controller/cronjobs/dal/internal/sql/queries.sql
+      - backend/controller/cronjobs/internal/sql/queries.sql
       - backend/controller/dal/internal/sql/async_queries.sql
     gen:
       go:
         <<: *gengo
-        out: "backend/controller/cronjobs/dal/internal/sql"
+        out: "backend/controller/cronjobs/internal/sql"
   - <<: *daldir
     queries: "internal/configuration/dal/internal/sql/queries.sql"
     gen:
